@@ -3,7 +3,6 @@ data handler and mapping resources for straightforward access.
 """
 import logging
 from pathlib import Path
-from typing import Optional
 
 from biocommons.seqrepo import SeqRepo
 
@@ -31,15 +30,6 @@ class CoolSeqTool:
     * ``self.alignment_mapper``: :py:class:`AlignmentMapper <cool_seq_tool.mappers.alignment.AlignmentMapper>`
     * ``self.mane_transcript``: :py:class:`ManeTranscript <cool_seq_tool.mappers.mane_transcript.ManeTranscript>`
     * ``self.ex_g_coords_mapper``: :py:class:`ExonGenomicCoordsMapper <cool_seq_tool.mappers.exon_genomic_coords.ExonGenomicCoordsMapper>`
-
-    Initialization with default resource locations is straightforward:
-
-    .. code-block:: pycon
-
-       >>> from cool_seq_tool.app import CoolSeqTool
-       >>> cst = CoolSeqTool()
-
-    See the :ref:`configuration <configuration>` section for more information.
     """
 
     def __init__(
@@ -48,10 +38,36 @@ class CoolSeqTool:
         lrg_refseqgene_path: Path | None = None,
         mane_data_path: Path | None = None,
         db_url: str = UTA_DB_URL,
-        sr: Optional[SeqRepo] = None,
+        sr: SeqRepo | None = None,
         force_local_files: bool = False,
     ) -> None:
-        """Initialize CoolSeqTool class
+        """Initialize CoolSeqTool class.
+
+        Initialization with default resource locations is straightforward:
+
+        >>> from cool_seq_tool.app import CoolSeqTool
+        >>> cst = CoolSeqTool()
+
+        By default, this will attempt to fetch the latest versions of static resources,
+        which means brief FTP and HTTPS requests to NCBI servers upon initialization.
+        To suppress this check and simply rely on the most recent locally-available
+        data:
+
+        >>> cst = CoolSeqTool(force_local_files=True)
+
+        Note that this will raise a FileNotFoundError if no locally-available data exists.
+
+        Paths to those files can also be explicitly passed to avoid checks as well:
+
+        >>> from pathlib import Path
+        >>> cst = CoolSeqTool(
+        ...     lrg_refseqgene_path=Path("lrg_refseqgene_20240625.tsv"),
+        ...     mane_data_path=Path("ncbi_mane_summary_1.3.txt")
+        ... )
+
+        If not passed explicit arguments, these locations can also be set via
+        environment variables. See the :ref:`configuration <configuration>` section of
+        the docs for more information.
 
         :param transcript_file_path: The path to ``transcript_mapping.tsv``
         :param lrg_refseqgene_path: The path to the LRG_RefSeqGene file
